added in chain specificity on quorum usage for different usages
