+++ DashSync/shared/Models/Chain/DSChain.m
@@ -92,6 +92,10 @@
 #define PROTOCOL_VERSION_LOCATION @"PROTOCOL_VERSION_LOCATION"
 #define DEFAULT_MIN_PROTOCOL_VERSION_LOCATION @"MIN_PROTOCOL_VERSION_LOCATION"
 
+#define ISLOCK_QUORUM_TYPE @"ISLOCK_QUORUM_TYPE"
+#define CHAINLOCK_QUORUM_TYPE @"CHAINLOCK_QUORUM_TYPE"
+#define PLATFORM_QUORUM_TYPE @"PLATFORM_QUORUM_TYPE"
+
 #define STANDARD_PORT_LOCATION @"STANDARD_PORT_LOCATION"
 #define JRPC_PORT_LOCATION @"JRPC_PORT_LOCATION"
 #define GRPC_PORT_LOCATION @"GRPC_PORT_LOCATION"
@@ -241,7 +245,7 @@ typedef NS_ENUM(uint16_t, DSBlockPosition)
     return self;
 }
 
-- (instancetype)initAsDevnetWithIdentifier:(NSString *)identifier checkpoints:(NSArray<DSCheckpoint *> *)checkpoints minimumDifficultyBlocks:(uint32_t)minimumDifficultyBlocks port:(uint32_t)port dapiJRPCPort:(uint32_t)dapiJRPCPort dapiGRPCPort:(uint32_t)dapiGRPCPort dpnsContractID:(UInt256)dpnsContractID dashpayContractID:(UInt256)dashpayContractID isTransient:(BOOL)isTransient {
+- (instancetype)initAsDevnetWithIdentifier:(NSString *)identifier checkpoints:(NSArray<DSCheckpoint *> *)checkpoints minimumDifficultyBlocks:(uint32_t)minimumDifficultyBlocks port:(uint32_t)port dapiJRPCPort:(uint32_t)dapiJRPCPort dapiGRPCPort:(uint32_t)dapiGRPCPort dpnsContractID:(UInt256)dpnsContractID dashpayContractID:(UInt256)dashpayContractID instantSendLockQuorumType:(DSLLMQType)instantSendLockQuorumsType chainLockQuorumType:(DSLLMQType)chainLockQuorumsType platformQuorumType:(DSLLMQType)platformQuorumType isTransient:(BOOL)isTransient {
     //for devnet the genesis checkpoint is really the second block
     if (!(self = [self initAsDevnetWithIdentifier:identifier checkpoints:checkpoints])) return nil;
     self.standardPort = port;
@@ -251,6 +255,9 @@ typedef NS_ENUM(uint16_t, DSBlockPosition)
     self.dashpayContractID = dashpayContractID;
     self.minimumDifficultyBlocks = minimumDifficultyBlocks;
     self.transient = isTransient;
+    self.quorumTypeForISLocks = instantSendLockQuorumsType;
+    self.quorumTypeForChainLocks = chainLockQuorumsType;
+    self.quorumTypeForPlatform = platformQuorumType;
     return self;
 }
 
@@ -351,7 +358,7 @@ static dispatch_once_t devnetToken = 0;
     return devnetChain;
 }
 
-+ (DSChain *)setUpDevnetWithIdentifier:(NSString *)identifier withCheckpoints:(NSArray<DSCheckpoint *> *)checkpointArray withMinimumDifficultyBlocks:(uint32_t)minimumDifficultyBlocks withDefaultPort:(uint32_t)port withDefaultDapiJRPCPort:(uint32_t)dapiJRPCPort withDefaultDapiGRPCPort:(uint32_t)dapiGRPCPort dpnsContractID:(UInt256)dpnsContractID dashpayContractID:(UInt256)dashpayContractID isTransient:(BOOL)isTransient {
++ (DSChain *)setUpDevnetWithIdentifier:(NSString *)identifier withCheckpoints:(NSArray<DSCheckpoint *> *)checkpointArray withMinimumDifficultyBlocks:(uint32_t)minimumDifficultyBlocks withDefaultPort:(uint32_t)port withDefaultDapiJRPCPort:(uint32_t)dapiJRPCPort withDefaultDapiGRPCPort:(uint32_t)dapiGRPCPort dpnsContractID:(UInt256)dpnsContractID dashpayContractID:(UInt256)dashpayContractID instantSendLockQuorumType:(DSLLMQType)instantSendLockQuorumsType chainLockQuorumType:(DSLLMQType)chainLockQuorumsType platformQuorumType:(DSLLMQType)platformQuorumType isTransient:(BOOL)isTransient {
     dispatch_once(&devnetToken, ^{
         _devnetDictionary = [NSMutableDictionary dictionary];
     });
@@ -359,7 +366,7 @@ static dispatch_once_t devnetToken = 0;
     __block BOOL inSetUp = FALSE;
     @synchronized(self) {
         if (![_devnetDictionary objectForKey:identifier]) {
-            devnetChain = [[DSChain alloc] initAsDevnetWithIdentifier:identifier checkpoints:checkpointArray minimumDifficultyBlocks:minimumDifficultyBlocks port:port dapiJRPCPort:dapiJRPCPort dapiGRPCPort:dapiGRPCPort dpnsContractID:dpnsContractID dashpayContractID:dashpayContractID isTransient:isTransient];
+            devnetChain = [[DSChain alloc] initAsDevnetWithIdentifier:identifier checkpoints:checkpointArray minimumDifficultyBlocks:minimumDifficultyBlocks port:port dapiJRPCPort:dapiJRPCPort dapiGRPCPort:dapiGRPCPort dpnsContractID:dpnsContractID dashpayContractID:dashpayContractID instantSendLockQuorumType:instantSendLockQuorumsType chainLockQuorumType:chainLockQuorumsType platformQuorumType:platformQuorumType isTransient:isTransient];
             _devnetDictionary[identifier] = devnetChain;
             inSetUp = TRUE;
         } else {
@@ -747,6 +754,96 @@ static dispatch_once_t devnetToken = 0;
             return;
         case DSChainType_DevNet: {
             setKeychainInt(protocolVersion, [NSString stringWithFormat:@"%@%@", self.devnetIdentifier, PROTOCOL_VERSION_LOCATION], NO);
+            break;
+        }
+    }
+}
+
+- (DSLLMQType)quorumTypeForISLocks {
+    switch ([self chainType]) {
+        case DSChainType_MainNet:
+            return DSLLMQType_50_60;
+        case DSChainType_TestNet:
+            return DSLLMQType_50_60;
+        case DSChainType_DevNet: {
+            NSError *error = nil;
+            DSLLMQType quorumType = (DSLLMQType)getKeychainInt([NSString stringWithFormat:@"%@%@", self.devnetIdentifier, ISLOCK_QUORUM_TYPE], &error);
+            if (!error && quorumType)
+                return quorumType;
+            else
+                return DSLLMQType_10_60;
+        }
+    }
+}
+
+- (void)setQuorumTypeForISLocks:(DSLLMQType)quorumTypeForISLocks {
+    switch ([self chainType]) {
+        case DSChainType_MainNet:
+            return;
+        case DSChainType_TestNet:
+            return;
+        case DSChainType_DevNet: {
+            setKeychainInt(quorumTypeForISLocks, [NSString stringWithFormat:@"%@%@", self.devnetIdentifier, ISLOCK_QUORUM_TYPE], NO);
+            break;
+        }
+    }
+}
+
+- (DSLLMQType)quorumTypeForChainLocks {
+    switch ([self chainType]) {
+        case DSChainType_MainNet:
+            return DSLLMQType_400_60;
+        case DSChainType_TestNet:
+            return DSLLMQType_50_60;
+        case DSChainType_DevNet: {
+            NSError *error = nil;
+            DSLLMQType quorumType = (DSLLMQType)getKeychainInt([NSString stringWithFormat:@"%@%@", self.devnetIdentifier, CHAINLOCK_QUORUM_TYPE], &error);
+            if (!error && quorumType)
+                return quorumType;
+            else
+                return DSLLMQType_10_60;
+        }
+    }
+}
+
+- (void)setQuorumTypeForChainLocks:(DSLLMQType)quorumTypeForChainLocks {
+    switch ([self chainType]) {
+        case DSChainType_MainNet:
+            return;
+        case DSChainType_TestNet:
+            return;
+        case DSChainType_DevNet: {
+            setKeychainInt(quorumTypeForChainLocks, [NSString stringWithFormat:@"%@%@", self.devnetIdentifier, CHAINLOCK_QUORUM_TYPE], NO);
+            break;
+        }
+    }
+}
+
+- (DSLLMQType)quorumTypeForPlatform {
+    switch ([self chainType]) {
+        case DSChainType_MainNet:
+            return DSLLMQType_100_67;
+        case DSChainType_TestNet:
+            return DSLLMQType_100_67;
+        case DSChainType_DevNet: {
+            NSError *error = nil;
+            DSLLMQType quorumType = (DSLLMQType)getKeychainInt([NSString stringWithFormat:@"%@%@", self.devnetIdentifier, PLATFORM_QUORUM_TYPE], &error);
+            if (!error && quorumType)
+                return quorumType;
+            else
+                return DSLLMQType_10_60;
+        }
+    }
+}
+
+- (void)setQuorumTypeForPlatform:(DSLLMQType)quorumTypeForPlatform {
+    switch ([self chainType]) {
+        case DSChainType_MainNet:
+            return;
+        case DSChainType_TestNet:
+            return;
+        case DSChainType_DevNet: {
+            setKeychainInt(quorumTypeForPlatform, [NSString stringWithFormat:@"%@%@", self.devnetIdentifier, PLATFORM_QUORUM_TYPE], NO);
             break;
         }
     }
