+++ DashSync/shared/Models/Chain/DSChainConstants.h
@@ -43,6 +43,10 @@
 #define MAX_VALID_MIN_PROTOCOL_VERSION 70219
 #define MIN_VALID_MIN_PROTOCOL_VERSION 70218
 
+#define DEVNET_ISLOCK_DEFAULT_QUORUM_TYPE 101
+#define DEVNET_CHAINLOCK_DEFAULT_QUORUM_TYPE 101
+#define DEVNET_PLATFORM_DEFAULT_QUORUM_TYPE 101
+
 #define DASH_MAGIC_NUMBER_TESTNET 0xffcae2ce
 #define DASH_MAGIC_NUMBER_MAINNET 0xbd6b0cbf
 #define DASH_MAGIC_NUMBER_DEVNET 0xceffcae2
