+++ DashSync/shared/Models/Managers/Chain Managers/DSChainsManager.h
@@ -42,9 +42,9 @@ FOUNDATION_EXPORT NSString *const DSChainsDidChangeNotification;
 
 - (DSChainManager *_Nullable)chainManagerForChain:(DSChain *)chain;
 
-- (void)updateDevnetChain:(DSChain *)chain forServiceLocations:(NSMutableOrderedSet<NSString *> *)serviceLocations withMinimumDifficultyBlocks:(uint32_t)minimumDifficultyBlocks standardPort:(uint32_t)standardPort dapiJRPCPort:(uint32_t)dapiJRPCPort dapiGRPCPort:(uint32_t)dapiGRPCPort dpnsContractID:(UInt256)dpnsContractID dashpayContractID:(UInt256)dashpayContractID protocolVersion:(uint32_t)protocolVersion minProtocolVersion:(uint32_t)minProtocolVersion sporkAddress:(NSString *_Nullable)sporkAddress sporkPrivateKey:(NSString *_Nullable)sporkPrivateKey;
+- (void)updateDevnetChain:(DSChain *)chain forServiceLocations:(NSMutableOrderedSet<NSString *> *)serviceLocations withMinimumDifficultyBlocks:(uint32_t)minimumDifficultyBlocks standardPort:(uint32_t)standardPort dapiJRPCPort:(uint32_t)dapiJRPCPort dapiGRPCPort:(uint32_t)dapiGRPCPort dpnsContractID:(UInt256)dpnsContractID dashpayContractID:(UInt256)dashpayContractID protocolVersion:(uint32_t)protocolVersion minProtocolVersion:(uint32_t)minProtocolVersion sporkAddress:(NSString *_Nullable)sporkAddress sporkPrivateKey:(NSString *_Nullable)sporkPrivateKey instantSendLockQuorumType:(DSLLMQType)instantSendLockQuorumType chainLockQuorumType:(DSLLMQType)chainLockQuorumType platformQuorumType:(DSLLMQType)platformQuorumType;
 
-- (DSChain *_Nullable)registerDevnetChainWithIdentifier:(NSString *)identifier forServiceLocations:(NSOrderedSet<NSString *> *)serviceLocations withMinimumDifficultyBlocks:(uint32_t)minimumDifficultyBlocks standardPort:(uint32_t)standardPort dapiJRPCPort:(uint32_t)dapiJRPCPort dapiGRPCPort:(uint32_t)dapiGRPCPort dpnsContractID:(UInt256)dpnsContractID dashpayContractID:(UInt256)dashpayContractID protocolVersion:(uint32_t)protocolVersion minProtocolVersion:(uint32_t)minProtocolVersion sporkAddress:(NSString *_Nullable)sporkAddress sporkPrivateKey:(NSString *_Nullable)sporkPrivateKey;
+- (DSChain *_Nullable)registerDevnetChainWithIdentifier:(NSString *)identifier forServiceLocations:(NSOrderedSet<NSString *> *)serviceLocations withMinimumDifficultyBlocks:(uint32_t)minimumDifficultyBlocks standardPort:(uint32_t)standardPort dapiJRPCPort:(uint32_t)dapiJRPCPort dapiGRPCPort:(uint32_t)dapiGRPCPort dpnsContractID:(UInt256)dpnsContractID dashpayContractID:(UInt256)dashpayContractID protocolVersion:(uint32_t)protocolVersion minProtocolVersion:(uint32_t)minProtocolVersion sporkAddress:(NSString *_Nullable)sporkAddress sporkPrivateKey:(NSString *_Nullable)sporkPrivateKey instantSendLockQuorumType:(DSLLMQType)instantSendLockQuorumType chainLockQuorumType:(DSLLMQType)chainLockQuorumType platformQuorumType:(DSLLMQType)platformQuorumType;
 
 - (void)removeDevnetChain:(DSChain *)chain;
